refactor(transport): Extract common test setup code to a function
Introduce `startRabbitMqContainer()` similar to `startArtemisContainer()`.

As now the container is installed for each test, it does not need to be
restarted in between tests anymore.
